requirements.txt
anthropic>=0.34.0
sentence-transformers>=3.0.0
fastapi>=0.112.0
uvicorn>=0.30.0
cryptography>=42.0.0
matplotlib>=3.9.0
scipy>=1.13.0
numpy>=1.26.0
python-dotenv>=1.0.0
